requirements.txt
numpy==1.17.4
dlib==19.22.1
Pillow==7.0.0
